Simplifying the rust port again, no need for additional bytes in the message to the port.
The roomba_port now inspects the first byte of the message sent from
elixir which is the roomba command and decides what to do. The only two
cases where the roomba_port does something different is when the first
byte is 0x8E or 0x01.

0x8E is the sensor command, and roomba_port ignores it instead of
sending it for now. Future release will make use of it.

0x01 is a pseudo command that turns on the roomba by calling the
function to pull the device line down. It's not in the SCI spec for the
roomba.